move query/configs to query_configs crate
