+++ query/configs/src/macros.rs
